requirements.txt
django
django-bootstrap-form
django-widget-tweaks
xlwt
django-qr-code
Pillow
xhtml2pdf
django-role-permissions
python-decouple
django-ninja
